Fix issues running emacs built from the native-comp branch (#2374)
The emacs native-comp branch provides "a modified Emacs capable of
compiling and running Emacs Lisp as native code in form of re-loadable
elf files" (https://akrl.sdf.org/gccemacs.html). Functions that are
natively-compiled do not return t to byte-code-function-p, but do return
t to subr-native-elisp-p, a function that is built-in to the native-comp
emacs.

The problem is that helm-symbol-name tries to call symbol-name on an
object if it is not functionp OR byte-code-function-p; when running with
native-comp, this now needs to include subr-native-elisp-p too. I've
also modified some other functions that use a similar pattern.